Build_external_kernelmodule.mk: Add support for kernel kit build module
The current version of build_module.sh has a limitation that it requires
a complete pre-compiled kernel source tree. This is not ideal as it
requires a complete kernel build available inside Android builds.

With KERNEL_KIT, only device/qcom/$(TARGET_BOARD_PLATFORM)-kernel
folder is needed, which could be committed or shared as a prebuilt. This
folder is much smaller and easier to share than the entire Kernel SI
output folder.

Kernel kit also has the added benefit of supporting incremental
compilation when rebuilding full kernel in Android workspace.

Change-Id: I7356d23d35ac6f57336e82bf6ff859dd873a9233